requirements.txt
Flask==2.3.3
requests>=2.28.0
python-dotenv>=1.0.0
huggingface-hub>=1.1.0
gunicorn>=20.1.0
